* Update ExtentSeries/ExtentField so that when a field is deleted, it      removes itself from the series, and so that a series can't be deleted      with any active fields.  This fixes an annoying bug that had to be      worked around in some other code.  Unfortunately it also showed a      remarkable number of cases where we deleted a series with live fields,      so those all had to be fixed.
DStoTextModule.H: make some of the set functions private so we can properly
  delete the underlying pointers.

ExtentField.H: move destructor out of header.

ExtentSeries.H: create a destructor, switch a few AssertAlways to INVARIANT

GeneralField.H: add a static function for deleting a vector of GeneralField *'s
  this pattern occurred a bunch of times

ExtentField.C: remove ourselves from the series on destruction

ExtentSeries.C: Don't let ourselves be destroyed if there are any
  active fields.  Also switch a bunch of AssertAlways to INVARIANTS.  Implement
  removeField

GeneralField.C: implement deleteFields

DStoTextModule.C: delete all the fields before the series

DataSeriesModule.C: remove int cast, not needed and was causing
  signed/unsigned warning.

dsextentindex.C: delete all the fields before the series

dsselect.C: delete the fields before the series

ellardnfs2ds.cpp: delete the fields before the series

test.C: test that the assertion is triggered when we try to
  incorrectly delete a series.  uses the new hooks in AssertBoost.

+++ src/module/DataSeriesModule.C
@@ -96,7 +96,7 @@ OutputModule::newRecord()
     INVARIANT(series.curExtent() == cur_extent,
 	      "usage error, someone else changed the series extent");
     INVARIANT(cur_extent != NULL, "called newRecord() after close()");
-    if ((int)(cur_extent->extentsize() + outputtype->fixedrecordsize()) > target_extent_size) {
+    if ((cur_extent->extentsize() + outputtype->fixedrecordsize()) > target_extent_size) {
 	double fixedsize = cur_extent->fixeddata.size();
 	double variablesize = cur_extent->variabledata.size();
 	double sumsize = fixedsize + variablesize;
